+++ AutomatedLabWorker/functions/VirtualMachines/New-LWHypervVM.ps1
@@ -29,11 +29,11 @@
     {
         $PSDefaultParameterValues['*:IsKickstart'] = $true
     }
-    if($Machine.OperatingSystemType -eq 'Linux' -and $Machine.LinuxType -eq 'Suse')
+    if ($Machine.OperatingSystemType -eq 'Linux' -and $Machine.LinuxType -eq 'Suse')
     {
         $PSDefaultParameterValues['*:IsAutoYast'] = $true
     }
-    if($Machine.OperatingSystemType -eq 'Linux' -and $Machine.LinuxType -eq 'Ubuntu')
+    if ($Machine.OperatingSystemType -eq 'Linux' -and $Machine.LinuxType -eq 'Ubuntu')
     {
         $PSDefaultParameterValues['*:IsCloudInit'] = $true
     }
@@ -124,9 +124,15 @@
             $ipSettings.Add('DnsDomain', $rootDomainName)
         }
 
-        if ($adapter.ConnectionSpecificDNSSuffix) { $ipSettings.Add('DnsDomain', $adapter.ConnectionSpecificDNSSuffix) }
+        if ($adapter.ConnectionSpecificDNSSuffix)
+        {
+            $ipSettings.Add('DnsDomain', $adapter.ConnectionSpecificDNSSuffix)
+        }
         $ipSettings.Add('UseDomainNameDevolution', (([string]($adapter.AppendParentSuffixes)) = 'true'))
-        if ($adapter.AppendDNSSuffixes)           { $ipSettings.Add('DNSSuffixSearchOrder', $adapter.AppendDNSSuffixes -join ',') }
+        if ($adapter.AppendDNSSuffixes)
+        {
+            $ipSettings.Add('DNSSuffixSearchOrder', $adapter.AppendDNSSuffixes -join ',')
+        }
         $ipSettings.Add('EnableAdapterDomainNameRegistration', ([string]($adapter.DnsSuffixInDnsRegistration)).ToLower())
         $ipSettings.Add('DisableDynamicUpdate', ([string](-not $adapter.RegisterInDNS)).ToLower())
 
@@ -155,7 +161,10 @@
 
     $Machine.NetworkAdapters = $adapters
 
-    if ($Machine.OperatingSystemType -eq 'Windows') {Add-UnattendedRenameNetworkAdapters}
+    if ($Machine.OperatingSystemType -eq 'Windows')
+    {
+        Add-UnattendedRenameNetworkAdapters
+    }
     #endregion network adapter settings
 
     Set-UnattendedComputerName -ComputerName $Machine.Name
@@ -245,7 +254,7 @@
 
     if ($Machine.Roles.Name -contains 'RootDC' -or
         $Machine.Roles.Name -contains 'FirstChildDC' -or
-    $Machine.Roles.Name -contains 'DC')
+        $Machine.Roles.Name -contains 'DC')
     {
         #machine will not be added to domain or workgroup
     }
@@ -265,7 +274,9 @@
                 Username = $domain.Administrator.UserName
                 Password = $domain.Administrator.Password
             }
-            if ($Machine.OrganizationalUnit) {$parameters['OrganizationalUnit'] = $machine.OrganizationalUnit}
+            if ($Machine.OrganizationalUnit) {
+                $parameters['OrganizationalUnit'] = $machine.OrganizationalUnit
+            }
 
             Set-UnattendedDomain @parameters
 
@@ -478,7 +489,7 @@
         ErrorAction = 'Stop'
     }
 
-    $vm = New-VM @vmParameter
+    $vm = Hyper-V\New-VM @vmParameter
 
     Set-LWHypervVMDescription -ComputerName $Machine.ResourceName -Hashtable @{
         CreatedBy = '{0} ({1})' -f $PSCmdlet.MyInvocation.MyCommand.Module.Name, $PSCmdlet.MyInvocation.MyCommand.Module.Version
@@ -569,7 +580,7 @@
     if ($Machine.HypervProperties.AutomaticStartAction) { $automaticStartAction = $Machine.HypervProperties.AutomaticStartAction }
     if ($Machine.HypervProperties.AutomaticStartDelay)  { $automaticStartDelay  = $Machine.HypervProperties.AutomaticStartDelay  }
     if ($Machine.HypervProperties.AutomaticStopAction)  { $automaticStopAction  = $Machine.HypervProperties.AutomaticStopAction  }
-    $vm | Set-VM -AutomaticStartAction $automaticStartAction -AutomaticStartDelay $automaticStartDelay -AutomaticStopAction $automaticStopAction
+    $vm | Hyper-V\Set-VM -AutomaticStartAction $automaticStartAction -AutomaticStartDelay $automaticStartDelay -AutomaticStopAction $automaticStopAction
 
     Write-ProgressIndicator
 
@@ -581,7 +592,7 @@
 
     if ( $Machine.OperatingSystemType -eq 'Windows')
     {
-        [void] (Mount-DiskImage -ImagePath $path)
+        [void](Mount-DiskImage -ImagePath $path)
         $VhdDisk = Get-DiskImage -ImagePath $path | Get-Disk
         $VhdPartition = Get-Partition -DiskNumber $VhdDisk.Number
 
@@ -802,7 +813,7 @@ Stop-Transcript
         $unattendXmlContent.Save("$VhdVolume\Unattend.xml")
         Write-PSFMessage "`tUnattended file copied to VM Disk '$vhdVolume\unattend.xml'"
         
-        [void] (Dismount-DiskImage -ImagePath $path)
+        [void](Dismount-DiskImage -ImagePath $path)
         Write-PSFMessage "`tdisk image dismounted"
     }    
 
@@ -828,9 +839,9 @@ Stop-Transcript
 
     $param = Sync-Parameter -Command (Get-Command Set-Vm) -Parameters $param
 
-    Set-VM -Name $Machine.ResourceName @param
+    Hyper-V\Set-VM -Name $Machine.ResourceName @param
 
-    Set-VM -Name $Machine.ResourceName -ProcessorCount $Machine.Processors
+    Hyper-V\Set-VM -Name $Machine.ResourceName -ProcessorCount $Machine.Processors
 
     if ($DisableIntegrationServices)
     {
@@ -845,7 +856,7 @@ Stop-Transcript
     Write-PSFMessage "Creating snapshot named '$($Machine.ResourceName) - post OS Installation'"
     if ($CreateCheckPoints)
     {
-        Checkpoint-VM -VM (Get-VM -Name $Machine.ResourceName) -SnapshotName 'Post OS Installation'
+        Hyper-V\Checkpoint-VM -VM (Hyper-V\Get-VM -Name $Machine.ResourceName) -SnapshotName 'Post OS Installation'
     }
 
     if ($Machine.Disks.Name)
@@ -858,6 +869,12 @@ Stop-Transcript
     }
 
     Write-ProgressIndicatorEnd
+
+    $writeVmConnectConfigFile = Get-LabConfigurationItem -Name VMConnectWriteConfigFile
+    if ($writeVmConnectConfigFile)
+    {
+        New-LWHypervVmConnectSettingsFile -VmName $Machine.Name
+    }
 
     Write-LogFunctionExit