+++ AutomatedLabWorker/functions/VirtualMachines/Start-LWHypervVM.ps1
@@ -16,7 +16,7 @@
         [switch]$NoNewLine
     )
 
-    if ($PreDelay) {
+    if ($PreDelaySeconds) {
         $job = Start-Job -Name 'Start-LWHypervVM - Pre Delay' -ScriptBlock { Start-Sleep -Seconds $Using:PreDelaySeconds }
         Wait-LWLabJob -Job $job -NoNewLine -ProgressIndicator $ProgressIndicator -Timeout 15 -NoDisplay
     }
@@ -27,7 +27,7 @@
 
         try
         {
-            Get-LWHypervVm -Name $Name.ResourceName | Start-VM -ErrorAction Stop
+            Get-LWHypervVM -Name $Name.ResourceName | Hyper-V\Start-VM -ErrorAction Stop
         }
         catch
         {
@@ -48,7 +48,7 @@
         }
     }
 
-    if ($PostDelay)
+    if ($PostDelaySeconds)
     {
         $job = Start-Job -Name 'Start-LWHypervVM - Post Delay' -ScriptBlock { Start-Sleep -Seconds $Using:PostDelaySeconds }
         Wait-LWLabJob -Job $job -NoNewLine:$NoNewLine -ProgressIndicator $ProgressIndicator -Timeout 15 -NoDisplay