+++ NewOrUpdateHelp.ps1
@@ -12,17 +12,17 @@
         'AutomatedLabWorker',
         'HostsFile',
         'AutomatedLabNotifications',
-        'AutomatedLab',
+        'AutomatedLabCore',
         'AutomatedLab.Recipe')
 )
 
-$location = $PSScriptRoot
+$buildFolder = if ($env:APPVEYOR_BUILD_FOLDER) { $env:APPVEYOR_BUILD_FOLDER } else { $PSScriptRoot }
 
 $outPath = foreach ($moduleName in $Module)
 {
-    $outputFolder = Join-Path $location -ChildPath "Help/$moduleName/en-us"
+    $outputFolder = Join-Path $buildFolder -ChildPath "Help/$moduleName/en-us"
     $outputFolder
-    Import-Module ./$moduleName -Force
+    Import-Module (Join-Path $buildFolder "publish/$moduleName") -Force
     if ($Create.IsPresent)
     {
         $null = New-MarkdownHelp -Module $moduleName -WithModulePage -OutputFolder $outputFolder -Force -AlphabeticParamsOrder
@@ -34,7 +34,7 @@ if (-not $Create.IsPresent)
     Update-MarkdownHelpModule -Path $outPath -RefreshModulePage -AlphabeticParamsOrder
 }
 
-foreach ($md in (Get-ChildItem -Filter *.md -Recurse -Path (Join-Path -Path $location -ChildPath Help)))
+foreach ($md in (Get-ChildItem -Filter *.md -Recurse -Path (Join-Path -Path $buildFolder -ChildPath Help)))
 {
     if (-not (Get-Command -ErrorAction SilentlyContinue -Name $md.BaseName)) { continue }
 
@@ -45,20 +45,19 @@ foreach ($md in (Get-ChildItem -Filter *.md -Recurse -Path (Join-Path -Path $loc
     $content | Set-Content -Path $md.FullName
 }
 
-$mkdocs = Join-Path -Path $location -ChildPath mkdocs.yml -Resolve -ErrorAction Stop
+$mkdocs = Join-Path -Path $buildFolder -ChildPath mkdocs.yml -Resolve -ErrorAction Stop
 $mkdocsContent = Get-Content -Raw -Path $mkdocs | ConvertFrom-Yaml
 
 # Update Sample Scripts help content
 $null = ($mkdocsContent.nav | Where-Object {$_.Keys -contains 'Sample scripts'})['Sample scripts'] = New-Object System.Collections.ArrayList
-foreach ($folder in (Get-ChildItem -Path (Join-Path -Path $location -ChildPath 'LabSources\SampleScripts') -Directory))
+foreach ($folder in (Get-ChildItem -Path (Join-Path -Path $buildFolder -ChildPath 'LabSources\SampleScripts') -Directory))
 {
-
     $folderObject = @{ $folder.Name = New-Object System.Collections.ArrayList}
     foreach ($sample in $folder.GetFiles('*.ps1'))
     {
         $mdRelativePathMkDocs = "Wiki/SampleScripts/$($folder.BaseName)/en-us/$($sample.BaseName).md"
         $mdRelativePath = "Help/Wiki/SampleScripts/$($folder.BaseName)/en-us/$($sample.BaseName).md"
-        $mdFullPath = Join-Path -Path $location -ChildPath $mdRelativePath
+        $mdFullPath = Join-Path -Path $buildFolder -ChildPath $mdRelativePath
         $scriptObject = @{
             $sample.Name = $mdRelativePathMkDocs
         }
@@ -87,18 +86,13 @@ $(Get-Content -Raw -Path $sample.FullName)
 # Update mkdocs.yml as part of a new help commit
 $null = ($mkdocsContent.nav | Where-Object {$_.Keys -contains 'Module help'})['Module help'] = New-Object System.Collections.ArrayList
 
-foreach ($moduleName in ($Module | Sort-Object))
-{
-    $moduleObject = @{$moduleName = New-Object System.Collections.ArrayList}
-    foreach ($command in (Get-Command -Module $moduleName))
-    {
-        $commandObject = @{
-            $command.Name = "$moduleName/en-us/$($command.Name).md"
-        }
-        $null = $moduleObject.$moduleName.Add($commandObject)
-    }
+$commands = Get-Command -Module $Module | Sort-Object Name -Unique
 
-    $null = ($mkdocsContent.nav | Where-Object {$_.Keys -contains 'Module help'})['Module help'].Add($moduleObject)
+foreach ($command in $commands)
+{
+    $null = ($mkdocsContent.nav | Where-Object {$_.Keys -contains 'Module help'})['Module help'].Add(@{
+        $command.Name = "$($command.Module.Name)/en-us/$($command.Name).md"
+    })
 }
 
 $mkdocsContent | ConvertTo-Yaml -OutFile $mkdocs -Force